Move user management to new User class
